+++ redux/types/employee-type.ts
@@ -1,4 +1,4 @@
-export const CHANGE_INPUT_VALUE = "CHANGE_INPUT_VALUE";
+export const CHANGE_INPUT_VALUE = "CHANGE_EMPLOYEE_INPUT_VALUE";
 export const CREATE_EMPLOYEE = "CREATE_EMPLOYEE";
 export const UPDATE_EMPLOYEE = "UPDATE_EMPLOYEE";
 export const GET_EMPLOYEE_LIST = "GET_EMPLOYEE_LIST";
